feat(racingcar): Implement the moving strategy
