+++ src/test/java/racingcar/MovingStrategy.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+@FunctionalInterface
+public interface MovingStrategy {
+    boolean movable();
+}
